requirements.txt
Django==3.2.3
generic==1.0.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
django-channels==0.7.0
django-channels-graphql-ws==0.8.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-auth==0.3.16
xdjango-graphql-jwt==0.3.0